+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.3.0
 
+* Fix: shouldn't auto play when seeking in pause state
 * keep screen on (disable screensaver) while playing video.
 
 ## 2.2.0